TEMP ENH(?): if preemptive AUTH failed without providing www-authenticate, retry without AUTH
In principle it seems that we should not do pre-emptive authentication
altogether and rely on the server to error out with 401 (possibly after some
redirects) and provide www-authenticate (on any schemas not only http_auth
using Basic but any of the standard ones) to act on.

For that to be implemented I think we should just RF the authentication and
authenticators logic to become such a dynamic one.

https://github.com/datalad/datalad/issues/5878 is the issue to pursue

and again decided against ;)